chore: factor out create/change fuel charge into component
